feat(common): Add support for using custom CAs with Volsync (#40000)
**Description**

Allows connecting volsync to an S3 API that uses a self-signed  cert.

Thiis is useful if you are self-hosting an S3 server and you are
exposing it via a self-signed certificate.

In my case this was needed because I am using the LAN IP of the server
and my Let's encrypt cert does not have it as one of its SANs. To avoid
making a LAN IP part of the cert history, I am instead using a
self-signed cert with the IP set up as SAN.

⚒️ Fixes  # <!--(issue)-->

**⚙️ Type of change**

- [X] ⚙️ Feature/App addition
- [ ] 🪛 Bugfix
- [ ] ⚠️ Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] 🔃 Refactor of current code
- [X] 📜 Documentation Changes

**🧪 How Has This Been Tested?**
<!--
Please describe the tests that you ran to verify your changes. Provide
instructions so we can reproduce. Please also list any relevant details
for your test configuration
-->

I rendered the authentik chart with the new common changes and deployed
it manually to my cluster.

What I have tested:

- Output is identical to `master` when `customCA` is unset
- When `customCA` is set to a non-empty string:
- `ReplicationDestination` and `ReplicationSource` resources correctly
render the `customCA` section
- An opaque secret with a single field (`ca.crt`) and name
`<chart-full-name>-volsync-ca-<credentials-name>` is correctly rendered
- The chart installs correctly with `kubectl apply -f
rendered-chart.yml`
- The volsync backup job prints "Using custom CA" and then succeeds in
creating a volume backup of the `blueprints` volume on my self-hosted S3
server with a self-signed cert
 
  
**📃 Notes:**
<!-- Please enter any other relevant information here -->

**✔️ Checklist:**

- [X] ⚖️ My code follows the style guidelines of this project
- [X] 👀 I have performed a self-review of my own code
- [ ] #️⃣ I have commented my code, particularly in hard-to-understand
areas
- [X] 📄 I have made changes to the documentation
- [X] 🧪 I have added tests to this description that prove my fix is
effective or that my feature works
- [X] ⬆️ I increased versions for any altered app according to semantic
versioning
- [X] I made sure the title starts with `feat(chart-name):`,
`fix(chart-name):`, `chore(chart-name):`, `docs(chart-name):` or
`fix(docs):`

**➕ App addition**

If this PR is an app addition please make sure you have done the
following.

- [ ] 🖼️ I have added an icon in the Chart's root directory called
`icon.png`

---

_Please don't blindly check all the boxes. Read them and only check
those that apply.
Those checkboxes are there for the reviewer to see what is this all
about and
the status of this PR with a quick glance._

---------

Signed-off-by: Alfred Göppel <43101280+alfi0812@users.noreply.github.com>
Co-authored-by: Alfred Göppel <43101280+alfi0812@users.noreply.github.com>