fix(guacamole) Guacamole Duo integration environment variables changes (#41223)
**Description**
<!--
Guacamole 1.6.0 with Duo two-factor authentication mechanism uses the
new Duo service v4, which requires different environment variables to be
defined
-->
⚒️ Fixes  # <!--(issue)-->

**⚙️ Type of change**

- [ ] ⚙️ Feature/App addition
- [X] 🪛 Bugfix
- [ ] ⚠️ Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] 🔃 Refactor of current code
- [ ] 📜 Documentation Changes

**🧪 How Has This Been Tested?**
<!--
- Created the environment variables for the main workload to validate
that these variables (according to guacamole documentation) work as
expected
- Ran the helm template to validate that the variables defined in the
standard way as overrides are defined in the guacamole pod
-->

**📃 Notes:**
<!-- Please enter any other relevant information here -->

**✔️ Checklist:**

- [X] ⚖️ My code follows the style guidelines of this project
- [X] 👀 I have performed a self-review of my own code
- [-] #️⃣ I have commented my code, particularly in hard-to-understand
areas
- [-] 📄 I have made changes to the documentation
- [-] 🧪 I have added tests to this description that prove my fix is
effective or that my feature works
- [-] ⬆️ I increased versions for any altered app according to semantic
versioning
- [X] I made sure the title starts with `feat(chart-name):`,
`fix(chart-name):`, `chore(chart-name):`, `docs(chart-name):` or
`fix(docs):`

**➕ App addition**

If this PR is an app addition please make sure you have done the
following.

- [-] 🖼️ I have added an icon in the Chart's root directory called
`icon.png`

---

_Please don't blindly check all the boxes. Read them and only check
those that apply.
Those checkboxes are there for the reviewer to see what is this all
about and
the status of this PR with a quick glance._

---------

Signed-off-by: Christian Tardif <christian.tardif@servinfo.ca>