feat(csi-addons-controller): Add csi-addons-controller (#41539)
**Description**

Adds a [CSI-Addons](https://github.com/csi-addons/spec/) controller
implementation for Kubernetes.


⚒️ Fixes  # <!--(issue)-->

**⚙️ Type of change**

- [x] ⚙️ Feature/App addition
- [ ] 🪛 Bugfix
- [ ] ⚠️ Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] 🔃 Refactor of current code
- [ ] 📜 Documentation Changes

**🧪 How Has This Been Tested?**
<!--
Please describe the tests that you ran to verify your changes. Provide
instructions so we can reproduce. Please also list any relevant details
for your test configuration
-->

Tested via deploying it to my cluster in 2 ways:

- As an app template (for testing before I created the chart)
- Rendered this chart with `helm template` and applied it to my cluster

In both cases, it deployed and worked as expected. 

What was tested:

- Works with rook-ceph(when the CSI-addons sidecar container is enabled)
- Successfully created a `ReclaimSpaceCronJob`s by annotating both a PVC
and a storage class with `reclaimspace.csiaddons.openshift.io/schedule:
"*/15 * * * *"`
- Watched the controller periodically creating `ReclaimSpaceJob`
resources from the `ReclaimSpaceCronJob` resources and updating them
with a successful status on job completion

**📃 Notes:**

I have not tested all
[CRDs](https://github.com/csi-addons/kubernetes-csi-addons/blob/main/deploy/controller/crds.yaml)
that come with the chart; however, they have been copied directly from
the [upstream
repository](https://github.com/csi-addons/kubernetes-csi-addons/tree/main),
along with the
[RBACs](https://github.com/csi-addons/kubernetes-csi-addons/blob/main/deploy/controller/rbac.yaml),
thus if a CRD does not work as expected, it is likely due to a bug with
the controller rather than the chart.

Also, I cannot find an icon for this project, so I have not added one to
the chart

**✔️ Checklist:**

- [x] ⚖️ My code follows the style guidelines of this project
- [x] 👀 I have performed a self-review of my own code
- [ ] #️⃣ I have commented my code, particularly in hard-to-understand
areas
- [ ] 📄 I have made changes to the documentation
- [x] 🧪 I have added tests to this description that prove my fix is
effective or that my feature works
- [ ] ⬆️ I increased versions for any altered app according to semantic
versioning
- [x] I made sure the title starts with `feat(chart-name):`,
`fix(chart-name):`, `chore(chart-name):`, `docs(chart-name):` or
`fix(docs):`

**➕ App addition**

If this PR is an app addition please make sure you have done the
following.

- [ ] 🖼️ I have added an icon in the Chart's root directory called
`icon.png`

---

_Please don't blindly check all the boxes. Read them and only check
those that apply.
Those checkboxes are there for the reviewer to see what is this all
about and
the status of this PR with a quick glance._