feat(grafana): Add dashboard auto-downloader (#41665)
**Description**
<!--
Please include a summary of the change and which issue is fixed. Please
also include relevant motivation and context. List any dependencies that
are required for this change.
-->

Currently, this chart allows provisioning of 3 predefined dashboards.

This is not ideal because:

- It is not flexible (only 3 statically defined in the chart dashboards
are supported)
- It bloats the chart unnecessarily
- Dashboards can get out of date quickly

To fix these shortcomings, I have added a curl init container that
executes a script which is generated during chart rendering based on the
values defined in the new map in the chart values
(`.Values.extraDashboards`). The map allows defining dashboards to be
provisioned both directly from `grafana.com` and via any URL that
returns a valid JSON Grafana dashboard file.

⚒️ Fixes  # <!--(issue)-->

**⚙️ Type of change**

- [x] ⚙️ Feature/App addition
- [ ] 🪛 Bugfix
- [ ] ⚠️ Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] 🔃 Refactor of current code
- [ ] 📜 Documentation Changes

**🧪 How Has This Been Tested?**
<!--
Please describe the tests that you ran to verify your changes. Provide
instructions so we can reproduce. Please also list any relevant details
for your test configuration
-->

- Rendered the chart manually with the values defined under
`.Values.extraDashboards`.
- Checked that the dashboards are present in Grafana after it starts

**📃 Notes:**

- The Grafana image technically already has `curl` as a command;
however, I chose to use a separate image just for the init container
because `curl` might be removed without notice from the Grafana image in
the future, and I do not want to depend on it being present.
- I have left the old way of deploying dashboards to prevent making this
a breaking change. After this is merged, I will open a new PR for
removing them (that will be marked breaking), and it can be merged when
appropriate.

**✔️ Checklist:**

- [x] ⚖️ My code follows the style guidelines of this project
- [x] 👀 I have performed a self-review of my own code
- [x] #️⃣ I have commented my code, particularly in hard-to-understand
areas
- [ ] 📄 I have made changes to the documentation
- [x] 🧪 I have added tests to this description that prove my fix is
effective or that my feature works
- [x] ⬆️ I increased versions for any altered app according to semantic
versioning
- [x] I made sure the title starts with `feat(chart-name):`,
`fix(chart-name):`, `chore(chart-name):`, `docs(chart-name):` or
`fix(docs):`

**➕ App addition**

If this PR is an app addition please make sure you have done the
following.

- [ ] 🖼️ I have added an icon in the Chart's root directory called
`icon.png`

---

_Please don't blindly check all the boxes. Read them and only check
those that apply.
Those checkboxes are there for the reviewer to see what is this all
about and
the status of this PR with a quick glance._

---------

Signed-off-by: Boemeltrein <130394941+Boemeltrein@users.noreply.github.com>
Signed-off-by: Kjeld Schouten <info@kjeldschouten.nl>
Co-authored-by: Boemeltrein <info@boemeltrein.nl>
Co-authored-by: Boemeltrein <130394941+Boemeltrein@users.noreply.github.com>
Co-authored-by: Kjeld Schouten <info@kjeldschouten.nl>