fix(wyoming-whisper): change persistence and securitycontext to remove root (#43142)
**Description**
Removes need for root privileges by expanding the persistence a bit
⚒️ Fixes  # <!--(issue)-->

**⚙️ Type of change**

- [ ] ⚙️ Feature/App addition
- [x] 🪛 Bugfix
- [ ] ⚠️ Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] 🔃 Refactor of current code
- [ ] 📜 Documentation Changes

**🧪 How Has This Been Tested?**
It works on my cluster:
[helmrelease.yaml](https://github.com/mgale456/ClusterTesting/blob/main/clusters/main/kubernetes/apps/whisper/app/helm-release.yaml)

**📃 Notes:**
huggingface has descriptions of their environment variables
[here](https://huggingface.co/docs/huggingface_hub/package_reference/environment_variables#hfhome).
There are more subdirectories under .cache/huggingface than just hub/,
so that's why I'm making this change (see HF_XET_CACHE or HF_TOKEN_PATH
as examples).

**✔️ Checklist:**

- [x] ⚖️ My code follows the style guidelines of this project
- [x] 👀 I have performed a self-review of my own code
- [ ] #️⃣ I have commented my code, particularly in hard-to-understand
areas
- [ ] 📄 I have made changes to the documentation
- [ ] 🧪 I have added tests to this description that prove my fix is
effective or that my feature works
- [x] ⬆️ I increased versions for any altered app according to semantic
versioning
- [x] I made sure the title starts with `feat(chart-name):`,
`fix(chart-name):`, `chore(chart-name):`, `docs(chart-name):` or
`fix(docs):`

**➕ App addition**

If this PR is an app addition please make sure you have done the
following.

- [ ] 🖼️ I have added an icon in the Chart's root directory called
`icon.png`

---

_Please don't blindly check all the boxes. Read them and only check
those that apply.
Those checkboxes are there for the reviewer to see what is this all
about and
the status of this PR with a quick glance._

---------

Signed-off-by: mgale456 <55673155+mgale456@users.noreply.github.com>
Co-authored-by: Alfred Göppel <43101280+alfi0812@users.noreply.github.com>